# reduce allocations by using values rather than pointers