--- a/api.go
+++ b/api.go
@@ -762,7 +762,7 @@ type RepoList struct {
 	Crashes int
 
 	// Minimal response to a List request. Returned when ListOptions.Minimal is true.
-	Minimal map[uint32]*MinimalRepoListEntry
+	Minimal map[uint32]MinimalRepoListEntry
 
 	// Stats response to a List request.
 	// This is the aggregate RepoStats of all repos matching the input query.