This reduces allocations even further by using values of MinimalRepoListEntry in the map rather than pointers. This reduces the number of allocations for decoding by another 76%. It slightly increases the allocated bytes because we copy the structs rather than pointers, and the structs are slightly larger.

Note that this is a change to the public API, so it is not an invisible change to the client. We would need to make a parallel change in Sourcegraph to consume the new minimal list type.

Stacked on #499. The benchmarks are relative to #499. 

```
❯ benchstat -alpha 0.05 /tmp/before.txt /tmp/after.txt
name                               old time/op    new time/op    delta
MinimalRepoListEncodings/slice-10    1.17ms ± 0%    1.19ms ± 1%   +1.52%  (p=0.000 n=10+10)
MinimalRepoListEncodings/map-10      1.48ms ± 0%    1.49ms ± 1%   +0.92%  (p=0.000 n=10+10)
RepoList_Encode-10                   55.5µs ± 0%    55.4µs ± 0%     ~     (p=0.063 n=9+9)
RepoList_Decode-10                   85.5µs ± 0%    74.7µs ± 1%  -12.68%  (p=0.000 n=10+9)

name                               old bytes      new bytes      delta
MinimalRepoListEncodings/slice-10     754kB ± 0%     754kB ± 0%     ~     (all equal)
MinimalRepoListEncodings/map-10       741kB ± 0%     741kB ± 0%     ~     (all equal)
RepoList_Encode-10                   50.9kB ± 0%    50.9kB ± 0%     ~     (all equal)

name                               old alloc/op   new alloc/op   delta
MinimalRepoListEncodings/slice-10    1.47kB ± 0%    1.49kB ± 0%   +1.27%  (p=0.000 n=9+9)
MinimalRepoListEncodings/map-10      53.8kB ± 0%    53.9kB ± 0%   +0.03%  (p=0.013 n=10+10)
RepoList_Encode-10                   57.3kB ± 0%    57.3kB ± 0%     ~     (all equal)
RepoList_Decode-10                    224kB ± 0%     249kB ± 0%  +11.32%  (p=0.000 n=10+10)

name                               old allocs/op  new allocs/op  delta
MinimalRepoListEncodings/slice-10      0.00           0.00          ~     (all equal)
MinimalRepoListEncodings/map-10       13.0k ± 0%     13.0k ± 0%     ~     (all equal)
RepoList_Encode-10                     1.00 ± 0%      1.00 ± 0%     ~     (all equal)
RepoList_Decode-10                    1.31k ± 0%     0.31k ± 0%  -76.16%  (p=0.000 n=10+10)
```